+++ test/mocha/promisify.js
@@ -1082,3 +1082,35 @@ describe("github 680", function() {
     });
 });
 
+describe("github 1063", function() {
+    specify("should not cause error when called with no arguments", function() {
+        return Promise.promisify(function(cb) {
+            cb();
+        }, { multiArgs: true})().then(function(values) {
+            assert(Array.isArray(values));
+            assert.strictEqual(values.length, 0);
+        });
+    })
+});
+
+describe("github 1023", function() {
+    specify("promisify triggers custom schedulers", function() {
+        var triggered = false;
+        var defaultScheduler = Promise.setScheduler(function(fn) {
+            triggered = true;
+            setTimeout(fn, 0);
+        });
+        var fnAsync = Promise.promisify(function(cb) {
+            setTimeout(function() {
+                cb(null, true);
+            }, 0);
+        });
+
+        return fnAsync().then(function(result) {
+            assert(result);
+            assert(triggered);
+        }).lastly(function() {
+            Promise.setScheduler(defaultScheduler);
+        });
+    });
+})