+++ docs/docs/changelog.md
@@ -3,6 +3,137 @@ id: changelog
 title: Changelog
 ---
 
+## 3.4.6 (2016-09-01)
+
+Bugfixes:
+
+- Fix [Promise.map](.) and [.map](.) not always calling the callback asynchronously \([#1148](.)\)
+
+## 3.4.5 (2016-08-31)
+
+Bugfixes:
+
+ - Fix unhandled error regression introduced in 3.4.3 [#1217](.)
+
+## 3.4.4 (2016-08-30)
+
+Bugfixes:
+
+ - Fix benchmark parallel in node 6 [#1165](.)
+ - Fix memory leak with Promise.each [#1057](.)
+ - Fix thenable passed to .return being evaluated too early [#1210](.)
+ - Fix "unhandledrejection" event not having .detail field when using DOM3 event listener api [#1209](.)
+ - Fix [Promise.join](.) not ensuring asynchronous callback [#1153](.)
+ - Fix domains leaking when synchronous error is thrown while a domain is active [#1125](.)
+
+
+
+## 3.4.3 (2016-08-25)
+
+Bugfixes:
+
+ - The "a promise was created in a handler but not returned from it" warning now highlights the file, line and column where the return statement is missing.
+  - The "a promise was created in a handler but not returned from it" warning now adds the bluebird API method used to create the non-returned promise at the top of the warning stack
+
+## 3.4.2 (2016-08-24)
+
+
+Bugfixes:
+
+ - Add missing link to unhandled warning docs \([#1205](.)\)
+ - Fix [Promise.delay](.) not having a long stack trace \([#1182](.)\)
+ - Fix false unhandled rejection when a rejected promise originating from one copy of bluebird is passed to another copy's [.return](.) or [.catchReturn](.) \([#1186](.)\)
+ - Fix Promise.resolve is not a function error \([#1192](.)\)
+ - Fix global events not being fired through DOM 3 API inside a worker \([#1190](.)\)
+ - Fix .cancel() not immediately marking a promise as being cancelled if it has cancellable parent \([#1187](.)\)
+ - Fix maximum callstack exceeded with [Promise.coroutine](.) \([#1170](.)\)
+
+## 3.4.1 (2016-06-17)
+
+Features:
+
+ - Added [Promise.getNewLibraryCopy](.)
+
+## 3.4.0 (2016-05-17)
+
+Features:
+
+ - Add `Promise.version` which tells the bluebird version as a string e.g. `"3.4.0"` ([#1042](.)).
+ - [.map](.), [Promise.map](.), [.filter](.) and [Promise.filter](.) now return rejected promise when inappropriate options argument is passed ([#1097](.)).
+
+Bugfixes:
+
+- Fix bug where callback to [.disposer](.) is not called if the resource is `null` ([#1099](.)).
+- Fix bug where assimilating thenable throws unexpectedly when using hostile host objects as thenables ([#1104](.)).
+
+## 3.3.5 (2016-04-12)
+
+Bugfixes:
+
+ - Fix then sometimes not being called on iOS/Firefox ([#1022](.)).
+ - Fix custom schedulers not being called when using promisified functions ([#1023](.)).
+ - Fix unexpected error being thrown when promisifed function is called with no arguments ([#1063](.)).
+
+## 3.3.4 (2016-03-07)
+
+Features:
+
+ - Warnings about created promises that are not returned are no longer given if the handler promise has not been chained. This should reduce the amount of false positives with this warning.
+
+## 3.3.3 (2016-02-25)
+
+Bugfixes:
+
+ - Fix stack overflow error when a promise returned by promisified function rejects early in a huge array when using [Promise.mapSeries](.) or [Promise.each](.)
+
+
+## 3.3.2 (2016-02-25)
+
+Bugfixes:
+
+ - Fix missing newline in stack trace reported by [.done()](.) ([#1020](.)).
+ - Detect deep circular resolutions
+
+## 3.3.1 (2016-02-13)
+
+Bugfixes:
+
+ - Fix crash when cancelling a [.tap()](.) handler promise ([#1006](.)).
+
+## 3.3.0 (2016-02-12)
+
+Features:
+
+ - Cancelling Promise returned from [Promise.delay()](.) and [.delay()](.) now calls `clearTimeout` ([#1000](.))
+ - Add [monitoring and lifecycle hooks](http://bluebirdjs.com/docs/features.html#promise-monitoring)
+ - Add `'warning'` hook for warnings ([#980](.))
+
+Bugfixes:
+
+ - Fix warnings for "promise was rejected with non-error" being output when promises are rejected with errors from different realm ([#990](.))
+
+
+## 3.2.2 (2016-02-05)
+
+Bugfixes:
+
+ - Make build script's output work without TTY
+
+## 3.2.1 (2016-02-01)
+
+Bugfixes:
+
+ - Revert monitoring feature due to crash in browser
+
+
+## 3.2.0 (2016-02-01)
+
+- Broken build
+
+## 3.1.5 (2016-01-26)
+
+Dummy release to trigger CDN update.
+
 ## 3.1.4 (2016-01-25)
 
 Bugfixes: